Refactor: Extract class/function definitions into separate modules
